+++ include/lib/util.hpp
@@ -37,6 +37,7 @@ inline void setbit(addr_t address, int bit) {
 inline void clearbit(addr_t address, int bit) {
   return set32(address, get32(address) & (~(1 << bit)));
 }
+#define SET_CLEAR_BIT(enable, addr, bit) (enable ? setbit : clearbit)(addr, bit)
 
 // ref:
 // https://github.com/torvalds/linux/blob/v6.8/arch/arm64/include/asm/sysreg.h#L1117-L1135
